+++ js/data.js
@@ -1,0 +1,24 @@
+export const avatars = [
+  'img/avatar-1.svg', 'img/avatar-2.svg', 'img/avatar-3.svg',
+  'img/avatar-4.svg', 'img/avatar-5.svg', 'img/avatar-6.svg'
+];
+
+export const names = ['Иван', 'Анна', 'Алексей', 'Ольга', 'Дмитрий', 'Елена'];
+
+export const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const descriptions = [
+  'Дорога',
+  'Горы',
+  'Море',
+  'Ужин',
+  'Кемпинг',
+  'Закат'
+];
